remove "c-unwind" unstable feature flag
The `extern "c-unwind"` ABI was stabilized in 1.71:
https://blog.rust-lang.org/2023/07/13/Rust-1.71.0.html#c-unwind-abi

There is still a nightly feature flag for this which covers the rest of that feature,
that being the guarantee that unwinding out of an `extern "C"` function safely aborts.
But this crate does not need that, so it's fine to remove the feature flag.